Ajust File Structure
