+++ src/assets/gatsby-astronaut.png
