+++ src/assets/gatsby-icon.png
