+++ src/assets/product-icon-institucional.png
