+++ src/components/SectionProducts/index.js
@@ -2,18 +2,10 @@ import React from "react"
 
 // Components
 import SectionContainer from "../shared/SectionContainer"
-import Card from "./Card"
+import Card from "./Card/"
 
 // Styled Components
-import styled from "styled-components"
-
-export const CardsContainer = styled.div`
-    display: grid;
-    grid-gap: 1rem;
-    grid-template-columns: repeat(2, 1fr);
-    justify-items: center;
-`;
-
+import {CardsContainer} from "./styled"
 
 const SectionProdutos = () => {
 
@@ -27,7 +19,6 @@ const SectionProdutos = () => {
         </CardsContainer>
     </SectionContainer>
     )
-
 }
 
 export default SectionProdutos
