+++ src/components/SectionProducts/styled.js
@@ -1,0 +1,9 @@
+// Styled Components
+import styled from "styled-components"
+
+export const CardsContainer = styled.div`
+    display: grid;
+    grid-gap: 1rem;
+    grid-template-columns: repeat(2, 1fr);
+    justify-items: center;
+`;
